repair state on start
