+++ libs/cosmos-sdk/server/start.go
@@ -64,8 +64,7 @@ func StartCmd(ctx *Context,
 	cdc *codec.Codec, appCreator AppCreator, appStop AppStop,
 	registerRoutesFn func(restServer *lcd.RestServer),
 	registerAppFlagFn func(cmd *cobra.Command),
-	appPreRun func(ctx *Context),
-	repairState func(ctx *Context)) *cobra.Command {
+	appPreRun func(ctx *Context)) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "start",
 		Short: "Run the full node",
@@ -98,10 +97,6 @@ which accepts a path for the resulting pprof file.
 			appPreRun(ctx)
 			// pruning options
 			_, err := GetPruningOptionsFromFlags()
-			//todo
-			if err == nil {
-				repairState(ctx)
-			}
 			return err
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
