+++ libs/cosmos-sdk/store/iavl/store.go
@@ -72,16 +72,12 @@ func LoadStoreWithInitialVersion(db dbm.DB, id types.CommitID, lazyLoading bool,
 	}, nil
 }
 
-// LoadStore returns an IAVL Store as a CommitKVStore setting its initialVersion
-// to the one given. Internally, it will load the store's version (id) from the
-// provided DB. An error is returned if the version fails to load.
-func LoadStoreWithInitialVersion123(db dbm.DB) (int64, error) {
-	//TODO
+func GetCommitVersion(db dbm.DB) (int64, error) {
 	tree, err := iavl.NewMutableTreeWithOpts(db, IavlCacheSize, &iavl.Options{InitialVersion: 0})
 	if err != nil {
 		return 0, err
 	}
-	return tree.LazyLoadVersion123(), nil
+	return tree.GetCommitVersion(), nil
 }
 
 // UnsafeNewStore returns a reference to a new IAVL Store with a given mutable
